# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"StatusCodes": {
				"enum": [
					100,
					101,
					102,
					103,
					200,
					201,
					202,
					203,
					204,
					205,
					206,
					207,
					300,
					301,
					302,
					303,
					304,
					305,
					307,
					308,
					400,
					401,
					402,
					403,
					404,
					405,
					406,
					407,
					408,
					409,
					410,
					411,
					412,
					413,
					414,
					415,
					416,
					417,
					418,
					419,
					420,
					421,
					422,
					423,
					424,
					426,
					428,
					429,
					431,
					451,
					500,
					501,
					502,
					503,
					504,
					505,
					507,
					511
				],
				"type": "number"
			},
			"UserResponse": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"firstName": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"username",
					"email",
					"lastName",
					"firstName",
					"id"
				],
				"type": "object"
			},
			"PaginationMetadata": {
				"properties": {
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"totalPages",
					"total",
					"limit",
					"page"
				],
				"type": "object"
			},
			"HttpPaginateResponse_UserResponse-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMetadata"
					}
				},
				"required": [
					"message",
					"code",
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllUserResponse": {
				"$ref": "#/components/schemas/HttpPaginateResponse_UserResponse-Array_"
			},
			"HttpResponse_UserResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserResponse"
					},
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"code",
					"message"
				],
				"type": "object"
			},
			"GetOneUserResponse": {
				"$ref": "#/components/schemas/HttpResponse_UserResponse_"
			},
			"PostUserResponse": {
				"$ref": "#/components/schemas/HttpResponse_UserResponse_"
			},
			"infer_typeofcreateUserSchema_": {
				"properties": {
					"lastName": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"email",
					"firstName",
					"password"
				],
				"type": "object"
			},
			"PostUserRequest": {
				"$ref": "#/components/schemas/infer_typeofcreateUserSchema_"
			},
			"PatchUserResponse": {
				"$ref": "#/components/schemas/HttpResponse_UserResponse_"
			},
			"infer_typeofupdateUserSchema_": {
				"properties": {
					"lastName": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"email",
					"firstName"
				],
				"type": "object"
			},
			"PatchUserRequest": {
				"$ref": "#/components/schemas/infer_typeofupdateUserSchema_"
			},
			"HttpResponse_undefined_": {
				"properties": {
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object"
			},
			"DeleteUserResponse": {
				"$ref": "#/components/schemas/HttpResponse_undefined_"
			},
			"RolePermissionResponse": {
				"properties": {
					"description": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"description",
					"name",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_RolePermissionResponse-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/RolePermissionResponse"
						},
						"type": "array"
					},
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"code",
					"message"
				],
				"type": "object"
			},
			"GetRolePermissionResponse": {
				"$ref": "#/components/schemas/HttpResponse_RolePermissionResponse-Array_"
			},
			"PutRolePermissionResponse": {
				"$ref": "#/components/schemas/HttpResponse_RolePermissionResponse-Array_"
			},
			"PermissionEnum": {
				"enum": [
					"user.read",
					"user.create",
					"user.update",
					"user.delete",
					"role.read",
					"role.create",
					"role.update",
					"role.delete",
					"booking.read",
					"booking.create",
					"booking.update",
					"booking.cancel"
				],
				"type": "string"
			},
			"infer_typeofupdateRolePermissionSchema_": {
				"properties": {
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/PermissionEnum"
						},
						"type": "array"
					}
				},
				"required": [
					"permissions"
				],
				"type": "object"
			},
			"PutRolePermissionRequest": {
				"$ref": "#/components/schemas/infer_typeofupdateRolePermissionSchema_"
			},
			"RoleResponse": {
				"properties": {
					"description": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"description",
					"name",
					"id"
				],
				"type": "object"
			},
			"HttpPaginateResponse_RoleResponse-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/RoleResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMetadata"
					}
				},
				"required": [
					"message",
					"code",
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllRoleResponse": {
				"$ref": "#/components/schemas/HttpPaginateResponse_RoleResponse-Array_"
			},
			"HttpResponse_RoleResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/RoleResponse"
					},
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"code",
					"message"
				],
				"type": "object"
			},
			"GetOneRoleResponse": {
				"$ref": "#/components/schemas/HttpResponse_RoleResponse_"
			},
			"PostRoleResponse": {
				"$ref": "#/components/schemas/HttpResponse_RoleResponse_"
			},
			"infer_typeofroleRequestSchema_": {
				"properties": {
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object"
			},
			"PostRoleRequest": {
				"$ref": "#/components/schemas/infer_typeofroleRequestSchema_"
			},
			"PatchRoleResponse": {
				"$ref": "#/components/schemas/HttpResponse_RoleResponse_"
			},
			"PatchRoleRequest": {
				"$ref": "#/components/schemas/infer_typeofroleRequestSchema_"
			},
			"DeleteRoleResponse": {
				"$ref": "#/components/schemas/HttpResponse_undefined_"
			},
			"GetCurrentUserResponse": {
				"$ref": "#/components/schemas/HttpResponse_UserResponse_"
			},
			"AuthenticatedUserRequest": {
				"properties": {
					"permissions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"username": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"permissions",
					"roles",
					"username",
					"lastName",
					"firstName",
					"email",
					"id"
				],
				"type": "object"
			},
			"PatchCurrentUserResponse": {
				"$ref": "#/components/schemas/HttpResponse_UserResponse_"
			},
			"PatchCurrentUserRequest": {
				"$ref": "#/components/schemas/infer_typeofupdateUserSchema_"
			},
			"AuthResponse": {
				"properties": {
					"refreshToken": {
						"type": "string"
					},
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken",
					"accessToken"
				],
				"type": "object"
			},
			"HttpResponse_AuthResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/AuthResponse"
					},
					"code": {
						"$ref": "#/components/schemas/StatusCodes"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"code",
					"message"
				],
				"type": "object"
			},
			"PostLoginResponse": {
				"$ref": "#/components/schemas/HttpResponse_AuthResponse_"
			},
			"infer_typeofloginSchema_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"PostLoginRequest": {
				"$ref": "#/components/schemas/infer_typeofloginSchema_"
			},
			"PostRefreshResponse": {
				"$ref": "#/components/schemas/HttpResponse_AuthResponse_"
			},
			"PostLogoutResponse": {
				"$ref": "#/components/schemas/HttpResponse_undefined_"
			},
			"PostRegisterResponse": {
				"$ref": "#/components/schemas/HttpResponse_UserResponse_"
			},
			"infer_typeofregisterSchema_": {
				"properties": {
					"lastName": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"confirmPassword",
					"password",
					"username",
					"firstName",
					"email"
				],
				"type": "object"
			},
			"PostRegisterRequest": {
				"$ref": "#/components/schemas/infer_typeofregisterSchema_"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "express-booking",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAllUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostUserRequest"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetOneUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatchUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatchUserRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/roles/{roleId}/permissions": {
			"get": {
				"operationId": "GetRoleWithAllPermissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRolePermissionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutRoleWithPermissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PutRolePermissionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutRolePermissionRequest"
							}
						}
					}
				}
			}
		},
		"/roles": {
			"get": {
				"operationId": "GetAllRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAllRoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostRoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostRoleRequest"
							}
						}
					}
				}
			}
		},
		"/roles/{roleId}": {
			"get": {
				"operationId": "GetOneRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetOneRoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatchRoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatchRoleRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteRoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Roles & Permissions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCurrentUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Current User"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatchCurrentUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Current User"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatchCurrentUserRequest"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostLoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostLoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostRefreshResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-Refresh-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostLogoutResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-Refresh-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostRegisterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostRegisterRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}